requirements.txt
# requirements.txt
sqlalchemy>=2.0
ib_insync==0.9.86
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pandas==2.2.3
pyarrow==23.0.1
python-dotenv==1.2.2
numpy==1.26.4
ta==0.11.0
yfinance==1.3.0